Move all connect/listen logic into send_recv.h.
This also fixes an issue where the batch server would use TCP even when STINGER_USE_TCP was OFF.
