+++ src/clients/streams/human_edge_generator/src/main.cpp
@@ -9,7 +9,6 @@
 
 #include "stinger_core/stinger.h"
 #include "stinger_core/xmalloc.h"
-#include "stinger_utils/stinger_sockets.h"
 #include "stinger_utils/timer.h"
 
 #include "human_edge_generator.h"
@@ -53,7 +52,7 @@ main(int argc, char *argv[])
   }
 
   /* start the connection */
-  int sock_handle = connect_to_batch_server (hostname, port);
+  int sock_handle = connect_to_server (hostname, port);
   if (sock_handle == -1) exit(-1);
 
   /* actually generate and send the batches */
