+++ src/clients/streams/netflow_stream/src/main.cpp
@@ -11,7 +11,6 @@
 
 #include "stinger_core/stinger.h"
 #include "stinger_core/xmalloc.h"
-#include "stinger_utils/stinger_sockets.h"
 #include "stinger_utils/timer.h"
 
 #include "netflow_stream.h"
@@ -80,7 +79,7 @@ int main(int argc, char *argv[])
 	LOG_D_A ("Sending batches every %d seconds", batch_time);
 
 	/* start the connection */
-	int sock_handle = connect_to_batch_server (hostname, port);
+	int sock_handle = connect_to_server (hostname, port);
 
 	/* actually generate and send the batches */
 	int64_t batch_num = 0;
