+++ src/clients/tools/json_rpc_server/src/mon_handling.cpp
@@ -158,50 +158,8 @@ mon_handler(void * args)
 int64_t
 mon_connect(int port, const char * host, const char * name)
 {
-#ifdef STINGER_USE_TCP
-  struct sockaddr_in sock_addr;
-  memset(&sock_addr, 0, sizeof(struct sockaddr_in));
-  struct hostent * hp = gethostbyname(host);
-
-  if(!hp) {
-    LOG_E_A("Error resolving host %s", host);
-    return -1;
-  }
-
-  int sock = socket(AF_INET, SOCK_STREAM, 0);
-  if(-1 == sock) {
-    LOG_E("Error opening socket");
-    return -1;
-  }
-
-  memcpy(&sock_addr.sin_addr.s_addr, hp->h_addr, hp->h_length);
-  sock_addr.sin_family = AF_INET;
-
-  if(port) {
-    sock_addr.sin_port = htons(port);
-  } else {
-    sock_addr.sin_port = htons(10103);
-  }
-
-  LOG_D_A("Socket open, connecting to host %s %s", host, inet_ntoa(**(struct in_addr **)hp->h_addr_list));
-#else
-  struct sockaddr_un sock_addr;
-  memset(&sock_addr, 0, sizeof(sock_addr));
-  int sock = socket(AF_UNIX, SOCK_STREAM, 0);
-  if(-1 == sock) {
-    LOG_E("Error opening socket");
-    return NULL;
-  }
-  strncpy(sock_addr.sun_path, "/tmp/stinger.sock", sizeof(sock_addr.sun_path)-1);
-  sock_addr.sun_family = AF_UNIX;
-
-  LOG_D_A("Socket open, connecting to host %s", host);
-#endif
-
-  if(-1 == connect(sock, (sockaddr *)&sock_addr, sizeof(sock_addr))) {
-    LOG_E("Error connecting socket");
-    return -1;
-  }
+  int sock = connect_to_server(host, port);
+  if (sock == -1) { return -1; }
 
   pthread_t mon_handler_thread;
   mon_handler_params_t * mon_params = (mon_handler_params_t *)xcalloc(1, sizeof(mon_handler_params_t));
