+++ src/server/src/batch_server.cpp
@@ -89,37 +89,13 @@ start_batch_server (void * args)
 
   int sock_handle, newsockfd;
   pthread_t garbage_thread_handle;
-  socklen_t clilen;
-  struct sockaddr_in serv_addr, cli_addr;
-  pid_t pid;
 
-  if (-1 == (sock_handle = socket(AF_INET, SOCK_STREAM, 0))) {
-    perror("Socket create failed.\n");
-    exit(-1);
-  }
+  sock_handle = listen_for_client(port_streams);
 
-  bzero ((char *) &serv_addr, sizeof(serv_addr));
-  serv_addr.sin_family = AF_INET;
-  serv_addr.sin_addr.s_addr = INADDR_ANY;
-  serv_addr.sin_port = htons(port_streams);
-
-  if (-1 == bind(sock_handle, (struct sockaddr *) &serv_addr, sizeof(serv_addr))) {
-    perror("Socket bind failed.\n");
-    exit(-1);
-  }
-
-  if (-1 == listen(sock_handle, 5)) {
-    perror("Socket listen failed.\n");
-    exit(-1);
-  }
-
-  clilen = sizeof(cli_addr);
-
-  LOG_V_A("STINGER server listening for input on port %d",
-      (int)port_streams);
+  LOG_V_A("STINGER batch server listening for input on port %d", port_streams);
 
   while (1) {
-    newsockfd = accept (sock_handle, (struct sockaddr *) &cli_addr, &clilen);
+    newsockfd = accept (sock_handle, NULL, NULL);
 
     if (newsockfd < 0) {
       perror("Accept new connection failed.\n");
